requirements.txt
fastapi[all]
pydantic
pyswisseph